requirements.txt
beautifulsoup4==4.12.3
httpx==0.27.0
polars==0.20.11
tqdm==4.66.2
pandas==2.2.1
python-dotenv==1.0.1
mysql-connector-python==8.3.0
